strtok_s
Hello everybody. In this project im trying to token an "inputed by the user" string. Then, count down all the white spaces in the added string. I was trying to accomplish this task by applying for loop , but nor the for loop and neither the "while" loop i putted there helped me achieve the result. I noticed that the problem is by:
1. number++ does not change its value at all
2. the white spaces counted from the pointer(string) are only for the first string input.
Any help would be highly appreciated. I prefer to use less of the new functions and try to code this task more of an old school/basic logic 
way. ThankS!